Update to latest (0.3) Trace Processor
Upgrading to the latest Trace Processor was mostly straightforward but
(for compatibility with .Net) the type for PIDs was changed from uint to
int, which required quite a few source-code updates.

Package update notes are here:

https://blogs.windows.com/windowsdeveloper/2020/01/28/traceprocessor-0-3-0/

+++ TraceProcessors/HeapSnapshotCompare/HeapSnapshotCompare.cs
@@ -50,7 +50,7 @@ namespace HeapSnapshotCompare
     // number of bytes and allocations outstanding.
     class SnapshotSummary
     {
-        public SnapshotSummary(Dictionary<ulong, AllocDetails> allocsByStackId, ulong pid)
+        public SnapshotSummary(Dictionary<ulong, AllocDetails> allocsByStackId, int pid)
         {
             allocsByStackId_ = allocsByStackId;
             pid_ = pid;
@@ -59,7 +59,7 @@ namespace HeapSnapshotCompare
         // column.
         public Dictionary<ulong, AllocDetails> allocsByStackId_;
         public Dictionary<string, long> hotStackFrames_ = null;
-        public ulong pid_;
+        public int pid_;
 
         public DataSize totalBytes_;
         public long allocCount_;
